chore: cleanup resource validators & utility code
