+++ Makefile
@@ -72,7 +72,7 @@ init:
 	go get ./...
 
 test:
-	go test -v ./...
+	go test ./...
 
 lint:
 	go run -modfile=tools/go.mod github.com/golangci/golangci-lint/cmd/golangci-lint run --fix
