+++ proxmoxtf/resource/validator/vm_test.go
@@ -4,7 +4,7 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
 
-package resource
+package validator
 
 import (
 	"testing"
@@ -33,7 +33,7 @@ func Test_getCPUTypeValidator(t *testing.T) {
 			t.Parallel()
 			require := require.New(t)
 
-			f := getCPUTypeValidator()
+			f := CPUType()
 			res := f(tt.value, nil)
 
 			if tt.valid {
@@ -41,41 +41,6 @@ func Test_getCPUTypeValidator(t *testing.T) {
 			} else {
 				require.NotEmpty(res, "validate: '%s'", tt.value)
 			}
-		})
-	}
-}
-
-func Test_parseImportIDWIthNodeName(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		name             string
-		value            string
-		valid            bool
-		expectedNodeName string
-		expectedID       string
-	}{
-		{"empty", "", false, "", ""},
-		{"missing slash", "invalid", false, "", ""},
-		{"valid", "host/id", true, "host", "id"},
-	}
-
-	for _, tt := range tests {
-		tt := tt
-		t.Run(tt.name, func(t *testing.T) {
-			t.Parallel()
-			require := require.New(t)
-
-			nodeName, id, err := parseImportIDWithNodeName(tt.value)
-
-			if !tt.valid {
-				require.Error(err)
-				return
-			}
-
-			require.Nil(err)
-			require.Equal(tt.expectedNodeName, nodeName)
-			require.Equal(tt.expectedID, id)
 		})
 	}
 }
