--- proxmoxtf/structure/tools.go
@@ -1,25 +1,0 @@
-/*
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/.
- */
-
-package structure
-
-import (
-	"fmt"
-
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
-)
-
-// MergeSchema merges the map[string]*schema.Schema from src into dst. Safety
-// against conflicts is enforced by panicking.
-func MergeSchema(dst, src map[string]*schema.Schema) {
-	for k, v := range src {
-		if _, ok := dst[k]; ok {
-			panic(fmt.Errorf("conflicting schema key: %s", k))
-		}
-
-		dst[k] = v
-	}
-}
